Create database Hotel

+++ III_C#_DB_Fundamentals/01.MS_SQL_Server/L01_Data_Definition/Database_Movies.sql
@@ -34,7 +34,7 @@ CREATE TABLE Movies(
 	[Length] DECIMAL(15,2),
 	GenreId INT NOT NULL,
 	CategoryId INT NOT NULL,
-	Rating INT,
+	Rating DECIMAL(10, 2),
 	Notes NVARCHAR(MAX),
 
 	CONSTRAINT PK_Movie PRIMARY KEY(Id),
